Adjust tests to correctly use the tmpdir fixture
tmpdir returns a LocalPath object (from py.path, path-like), which is
the object-oriented interface for the os.path library. The problems
occur when these objects are used within os.path; this, however is
solved in the previous commits by adjust the Multiverse and the
tools.read_yml and write_yml functions to case to strings.