Adjust dummy model to be hierarchically decoupled
This entails: initialising the RNG, the HDF5 file, and the Configuration
at the top level (i.e., in the main function) and adjusting the Dummy
Model class constructor to select the corresponding entries when the
model is initialised.

This step is required here because otherwise it is not possible to test
the adjust model configuration.