Add the extract_ia_matrix function (outcommented)
A function that extracts the correct interaction matrix from the configuration is implemented.
It does not work yet because of lacking functionality of the `Config` class that is provided by a YAML::Node.
This problem should be solved with issue #136. Untill then, the function is commented out and returns (and requires) an explicitely given interaction matrix in the config.