Use function to invert nb_mode_map
This is better than specifying the map twice, once from string to enum
and one in reverse.
Ideally, this would all happen at compile time, but seems only be doable
with a lot of overkill template gedöns or precompiler macros; not keen
on doing that, as the performance is not critical in this case.