Kaching! Use uniform rounding for transformation back into the space
This is the easiest and safest way to do it. And it does not require a
distinction between positive and negate relative position!
The crucial part is that a rounding method that always rounds into the
same direction needs to be used. The int constructor / cast does not do
so, but std::floor and std::ceil do!