Pass config parameters through select_neighborhood interface
The neighborhood parameters cannot be set during construction because
they need to be coupled to the neighborhood mode that is selected, which
is changed by the CellManager's and the Grid's select_neighborhood
methods.
These now pass a DataIO::Config node along, which can be evaluated in
the get_nb_func method, and extract the required parameters for a
certain neighborhood mode.