Remove check for distance in vonNeumann NBFunc; minor tweaks
Regarding the check: The function still works for distance 1, but is
just not as efficient. However, the check for the distance is made
elsewhere and the more specialized function is selected there; thus, it
is not necessary to check again.
This also has the advantage of reducing indentation level by one. :)

Other tweaks:
   - In the calculation of expected number of neighbors, the static
   class member `dim` was shadowed; renamed to d
   - Code layout and consistency improvements
   - More consistent error messages