Define CellManager::RuleFunc and improve apply_rule
The RuleFunc typedef is a convenience definition in the CellManager as
all types are available there.

apply_rule now takes a const std::shared_ptr<Cell>& and a const Rule&,
which avoids copying of the function objects on each apply_rule call.
Tests were adjusted to show that this also works with temporary objects.