Handle pip show cases where the package is NOT installed
The detection of whether a package is installed or not used to work via
the version being specified in pip freeze; with pip show, there is the
additional case where a version is not available, e.g. because it is not
set in that package's setup.py. Thus, python_find_package needs to
handle an additional case, where a package was found, but no version is
specified. If in such a case, a version was required, it is not marked
as found; if no version was given, it is still marked as found.