Rework CI/CD and docker images
There no longer is the need for a prep stage, because there no longer is
DUNE that might need to be updated. Now, rather than updating the base
image, the base image is just built anew each time there is a push to
master - that way, the number of layers stays fixed.