Clean up CLI and better modularise model registry
The CLI was mainly improved by cleaning up the formatting and grouping
arguments together such that the --help commands are easier to grasp.

The model registry was improved by adding path parsing to the bundles
and removing redundant methods; there should now be a clear way on how
to access model information and how to register new models.