Let ExternalPlotCreator use custom plot_modules_path config entry
Instead of adding the parent paths to the sys.path, this now adds the
paths as specified to the sys.path, thus making the contained folders
available for import. That allows to have multiple additional plot
modules (for many Utopia installations) without overwriting previously
existing ones.
They all remain available under the model_plots. prefix, which is
crucial for backwards compatibility.

This approach still needs some polishing; it's unclear how robust it is.