Encapsulate nb_funcs into generator method to allow capture
This makes it possible to capture any parameters that are related to the
computation of the neighborhood (e.g. the distance) and at the same time
allows to move some computations outside of the neighborhood function,
e.g. the calculation of the neighborhood size.

With this, only some neighborhood functions are stored as members of the
Grid and Grid-derived classes. Those that need the capture are now
encapsulated in a generator method that returns the NBFunc<Grid>.

The changes introduced here will not only bring about some minor
speed improvements, but also provides a more elegant solution than
storing the parameters of the neighborhood function as separate members
of the Grid class. Instead, they are stored as a Config node, which is
potentially empty.

This will make #16 (multiple neighborhoods) much easier to implement.