Add a description when to use which way to configure utopia.
Move the config madness Faq part to the frontend and reworked a lot
of it. Namely inserted an overview of the hierarchy and make the text
more easy to read plus correct some typos

Change the headline of the config madness chapter

Remove the leading _ from references

Add a description when to use which way to configure utopia.

Shifted the config madness Faq part to the frontend and reworked a lot
of it. Namely inserted an overview of the hierarchy and make the text
more easy to read plus correct some typos

Changed the headline of the config madness chapter

removed the leading _ from references