Remove compiler warning in structured bindungs unused variable
Unfortunately, in the current compiler versions it is not possible
to mute a warning by adding the [[maybe_unused]] argument in front of
the unpacked tuple if only a subset of entities is not used.
So, to get rid of the warning, structured binding was not used, instead
the iterator pair was stored in one it variable.