Adapt and extend utopya.cltools tests
This includes tests for all lines that can be reasonably well tested via
pytest. The tests cannot really test the copying of the files; that
would be quite a mess.

This also adds a minor feature to allow skipping certain file extensions
and it improves the error handling upon failing to read a file.