Fix nonzero_exit_handling param being ignored
This was due to the Model class setting 'raise' as the default value,
while it should actually be configurable from the meta configuration.

As people now got used to 'raise' being the default, this commit sets it
in the proper place: the base configuration, which used to contain as
parameter the 'warn_all' setting.