Migrate model_test.cc to use Boost Test Framework
This also changes the `monitor_emit_interval` parameter to 2s in order
to (hopefully) mitigate a race condition in this test.

Also, it replaces all variables used in the tests by temporaries in
order to avoid any confusion of which values are currently tested
against.

The Boost.Test capabilities further remove the need to have a
`compare_containers` method, which is thus deleted from `model_test.hh`.