Allow proper CRTP for derivation from Entity
The lack of this led to the CustomLinks type being specialized with the
Entity, and not with the type of the *derived* class.
By additionally templating Entity with the `Derived` type, CRTP is
complete, thus allowing to spezialize the CustomLinks type with the type
of the *derived* class.

The changes here are mostly internal and should not require any further
adjustments in models, unless custom `Entity`-derived types are defined
somewhere.