Rework ccees/utopia docker image to include source files
This simplifies the dockerfile in various ways.

As the source files may remain inside the docker image, we no longer
need to recursively copy them over from the build directory; everything
can remain in place as it is.

Furthermore, with dantro and paramspace now publicly accessible, we do
not need read-only SSH access to the GitLab projects any more. With this
step gone, there is no reason left to divide the build into a
multi-stage build; previously, this was done to keep the SSH key safe.