Add explicit git checkout and remote set-url; copy cmake config
With this commit, it also becomes possible to specify the docker image
tag used in the `ci_test` stage via a CI/CD variable and the GitLab
CI/CD web interface.

Note: git is not available in the DinD docker:stable image, thus it is
not possible to do this inside the CI job.