Adjust Utopia::Model constructor to allow custom configuration
This adjusts the Utopia::Model constructor to pass writer args as an
std::tuple of arguments, rather than a parameter pack that is later
packed into a tuple.

Furthermore, an additional argument `DataIO::Config& custom_cfg` is
added, which specifies a custom configuration node to use instead of the
one inherited from the parent model. This can be useful when dynamically
instantiating models.

The `custom_cfg` argument is added in front as it is the common
functionality to all models, while the `writer_args` are only used when
using the `DataManager`.

Also declares writer_args as [[maybe_unused]] as it may be unused
depending on WriteMode.