In Multiverse, decouple adding tasks and running
Additionally, this adds the private `_shared_worker_manager` argument,
which EXPERIMENTALLY allows to share a WorkerManager between Multiverse
instances. Subsequently, multiple Multiverse runs can be chained
together, using the scheduling capabilities of the WorkerManager.

Through the new modularisation of the Multiverse, it is possible to add
simulation tasks without locking the WorkerManager's task list and
starting to work. This way, it becomes possible to first aggregate a
number of tasks and then, after all tasks were added, let the
WorkerManager start to work.

This change is EXPERIMENTAL because it does not provide a smooth user
experience. This is mostly due to the Reporter instance, which also has
to be shared among Multiverses. As the Reporter implicitly assumes that
only tasks from a single Multiverse were added to the WorkerManager, it
will display *wrong* information: it shows the _aggregated_ information
instead of that for each Multiverse. Furthermore, it writes that
information to the first Multiverse's output directory, and not
individually to all Multiverses.

Given that the option to use a shared WorkerManager is clearly marked as
experimental, these limitations seem to be acceptable. The alternative
would be a thorough rework of the concepts of a Multiverse, which does
not seem feasible nor desirable at this point. (We have always assumed
to have a Multiverse as the "largest unit"...)