Fix missing recursive update when building meta-configuration
The bug occured when the base or user configuration contained entries
that were specific to a single model; while this is not something that
is advised to be done, it is unexpected that there would *not* be a
recursive update there (while we do recursive updates literally
everywhere else). On the off chance that someone is putting
model-specific information into the user configuration, it should thus
not fail silently but do what is expected. The other option would be to
prohibit this, but that seems like an unnecessary restriction ...