Let MonitorManager write nested monitor entries
This brings consistent handling of model nesting to the monitoring
feature. Up unto this point, monitor entries were written to keys of the
form `<ModelName>.<key>`, which would require all models to have unique
names, an assumption we make nowhere and don't check against.

Staying true to the nestedness of the *input* model configuration, the
monitor output can use the same approach: Write monitor output as a
mapping of the following form:

    !!map {time: 123, progress: 0.5, MyModel: {foo: 123, bar="yes!"}}

This uses the `recursive_setitem` functionality and the (newly added)
`full_name` of each model: basically a concatenation of model names of
the form:

    .MyModel.MySubModel.MySubSubModel

... and so on. When setting a monitor entry, the path is walked along,
splitting at each `.` delimiter.

The performance impact of this should be negligible; walking along a few
keys is not too expensive and the monitor is only called on the order of
seconds. There might even be a benefit, given that the string sent over
stdout is shorter -- and also shorter to parse on frontend-side.

There is a visual benefit as well: entries are much easier to read,
because the model name is not prepended to each monitor key.