Improve doxygen docs and address some warnings
Warnings were due to missing doxygen XML output for the
Utopia::get_as functions and Utopia::SelectionMode enum.
Workaround was to refer to the doxygen group instead when referencing it
from the sphinx docs (via breathe).

Fix broken links in docs:

Some links are not actually broken but (now) redirect to sites that do
not allow (or redirect to sites that do not allow) the Sphinx linkcheck
user agent and respond with 403.